Fixed logic of adding and deleting orders, caught some exceptions(try-catch) and made proper Response entities; also there shouldn't be any problem of adding new items to orders; only owners of the orders can add, delete, or do any other kind of manipulations to the orders. Next step is to go through the status setting methods -> all statuses should be updated after making any manipulations to the orders Think about the front end of the application too -> maybe I'll need to implement Vaadin full-stack application framework to create a good looking modern interface for myCourier application.
Do I have to use Heroku or Do have to user Heroku only for server side and place front-end on a free hosting?

+++ src/main/java/com/example/demo/controller/OrderController.java
@@ -1,19 +1,19 @@
 package com.example.demo.controller;
 
+import com.example.demo.enums.CourierStatus;
 import com.example.demo.model.*;
 import com.example.demo.repository.CourierRepository;
+import com.example.demo.repository.OrderRepository;
 import com.example.demo.repository.UserRepository;
 import com.example.demo.service.OrderService;
 import com.example.demo.service.UserRolesService;
 import com.example.demo.service.UserService;
-import org.apache.tomcat.util.net.openssl.ciphers.Authentication;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.security.access.annotation.Secured;
-import org.springframework.security.access.prepost.PostAuthorize;
-import org.springframework.security.access.prepost.PreAuthorize;
 import org.springframework.web.bind.annotation.*;
 
 import java.security.Principal;
+import java.util.NoSuchElementException;
 
 @CrossOrigin(origins = "*", allowedHeaders = "*")
 @RestController
@@ -24,48 +24,78 @@ public class OrderController {
     @Autowired
     UserRolesService userRolesService;
     @Autowired
+    OrderRepository orderRepository;
+    @Autowired
     UserRepository userRepository;
     @Autowired
     CourierRepository courierRepository;
     @Autowired
     UserService userService;
     //-------------------------------
-    @RequestMapping(value = "/username", method = RequestMethod.GET)
-    @ResponseBody
-    public User currentUserName(Principal principal) {
-        return userRepository.findByUsername(principal.getName());
-    }
+
     //-------------------------------
+    //tested
     @Secured("ROLE_ADMIN")
     @GetMapping(path = "/getAll")
     public Response getAll(){
         return new Response(true, "All orders", orderService.getAllOrders());
     }
+    //tested - needs one more final test
     @Secured("ROLE_ADMIN")
     @GetMapping(path = "/getById/{id}")
-    public Response getById(@PathVariable Long id){
-        return new Response(true,"Order with id = " + id, orderService.getOrderById(id));
+    public Response getById(Principal principal, @PathVariable Long id){
+        try {
+            return new Response(true, "Order with id = " + id, orderService.getOrderById(id));
+        }
+        catch (NoSuchElementException exception){
+            System.out.println(exception.getMessage());
+            return new Response(false, "Order with id = "+id+ " doesn't exist", exception.getMessage());
+        }
     }
+    //tested - needs one more final test
     @Secured("ROLE_USER")
     @PostMapping(path = "/create")
     public Response createOrder(Principal principal, @RequestBody Order order){
         order.setOrderedBy(userRepository.findByUsername(principal.getName()));
         return new Response(true, "Order created successfully", orderService.createOrder(order));
     }
+    //tested - needs one more final test
     @PutMapping(path = "/addItemToOrder/{orderId}")
-    public Response addItemToOrder(@PathVariable Long orderId, ItemWrapper itemWrapper){
-        return new Response(true, "New items were added to order", orderService.addNewItemsToOrder(itemWrapper,orderId));
+    public Response addItemToOrder(Principal principal, @PathVariable Long orderId, @RequestBody ItemWrapper itemWrapper){
+        if(!orderRepository.existsById(orderId)){
+            return new Response(false, "Such order doesn't exits", null);
+        }
+        if(orderRepository.belongsTo(orderId,principal.getName()).equals(1)){
+            return new Response(true, "New items were added to order", orderService.addNewItemsToOrder(itemWrapper,orderId));
+        }
+        else {
+            return new Response(false, "Order with id = " + orderId + " doesn't belong to the current user", null);
+        }
     }
     //tested - need's one more test
     @Secured("ROLE_COURIER")
     @PutMapping(path = "/takeOrder/{id}")
     public Response takeOrder(Principal principal, @PathVariable Long id){
+        if(!orderRepository.existsById(id)){
+            return new Response(false, "Such order doesn't exist",null);
+        }
+        if(courierRepository.findByUsername(principal.getName()).getStatus().equals(CourierStatus.BANNED)){
+            return new Response(false, "Current courier has the status = BANNED, which doesn't allow to take any orders",null);
+        }
         return new Response(true, "Order taken by courier with username = " + principal.getName() , orderService.takeOrder(id, courierRepository.findByUsername(principal.getName())));
     }
-    @PostAuthorize("#id == returnObject")
+    //tested
     @DeleteMapping(path = "/delete/{id}")
-    public Response deleteOrder(@PathVariable Long id){
-        orderService.deleteOrder(id);
-        return new Response(true, "Order with id = " + id + " deleted!!", null);
+    public Response deleteOrder(Principal principal, @PathVariable Long id){
+        if(!orderRepository.existsById(id)){
+            return new Response(false, "Such order doensn't exist", null);
+        }
+        if(orderRepository.belongsTo(id,principal.getName()).equals(1)){
+            orderService.deleteOrder(id);
+            return new Response(true, "Order with id = " + id + " deleted!!", null);
+        }
+        else{
+            return new Response(false, "Order with id = " + id + " doesn't belong to current user", null);
+        }
     }
 }
